refactor: 💡 minor adjustments
