+++ apps/e-learning-app/src/app/components/contentGrid/ContentGridStyles.js
@@ -1,0 +1,6 @@
+const contentGridStyles = {
+  padding: '40px 10px 40px 0',
+  marginLeft: '17vw',
+};
+
+export default contentGridStyles;
