+++ apps/e-learning-app/src/app/components/moduleContent/ModuleContent.js
@@ -1,12 +1,17 @@
-import ContentGrid from './ContentGrid';
-import Sidebar from './moduleSidebar/Sidebar';
-import { allVideos } from './MockData';
+import ContentGrid from '../contentGrid/ContentGrid';
+import Sidebar from '../moduleSidebar/Sidebar';
+import { allVideos } from '../MockData';
+
 /**
  * MODULE PAGE
  *
  * ModuleContent renders:
  * - the Sidebar component (from ./moduleSidebar/Sidebar.js)
  * - the ContentGrid component (from ./ContentGrid.js)
+ * 
+ * Props:
+ * 
+ * - title
  */
 
 const ModuleContent = (props) => {
@@ -14,7 +19,7 @@ const ModuleContent = (props) => {
   return (
     <>
       <Sidebar title={title} />
-      <ContentGrid type="normal" label="Recommended" videos={allVideos} />
+      <ContentGrid videos={allVideos} />
     </>
   );
 };
