+++ apps/e-learning-app/src/app/components/moduleVideo/Video.js
@@ -12,7 +12,7 @@ import style from './VideoStyle';
  * Props:
  * - thumbnail: string (the path of the photo)
  * - title: string (the name displayed under the photo)
- * - videoLink: string (the link of the video's coresponding page)
+ * - videoLink: string (the link of the video's corresponding page)
  * - channel: string (the link to the user that created the video)
  * - length: int (number of seconds the video takes)
  *
@@ -24,10 +24,10 @@ const Video = (props) => {
 
   const getLengthString = (length) => {
     // 208 -> 03:28
-    var sec_num = parseInt(length);
-    var hours = Math.floor(sec_num / 3600);
-    var minutes = Math.floor(sec_num / 60) % 60;
-    var seconds = sec_num % 60;
+    const sec_num = parseInt(length);
+    const hours = Math.floor(sec_num / 3600);
+    const minutes = Math.floor(sec_num / 60) % 60;
+    const seconds = sec_num % 60;
 
     return [hours, minutes, seconds] // 0, 3, 28
       .map((v) => (v < 10 ? '0' + v : v)) // 00, 03, 28
