+++ apps/e-learning-app/src/app/pages/modulePage/ModulePage.js
@@ -1,4 +1,5 @@
-import ModuleContent from '../components/ModuleContent';
+import ModuleContent from '../../components/moduleContent/ModuleContent';
+import style from './ModulePageStyles';
 
 /**
  * MODULE PAGE
@@ -8,13 +9,8 @@ import ModuleContent from '../components/ModuleContent';
  */
 
 const ModulePage = () => {
-  const content_style = {
-    padding: '10px 5.2vw',
-    display: 'flex',
-    width: '100%',
-  };
   return (
-    <div style={content_style}>
+    <div style={style}>
       <ModuleContent title="II. Studiu de caz" />
     </div>
   );
