+++ apps/e-learning-app/src/app/pages/modulePage/ModulePageStyles.js
@@ -1,0 +1,7 @@
+const modulePageStyles = {
+  padding: '10px 5.2vw',
+  display: 'flex',
+  width: '100%',
+};
+
+export default modulePageStyles;
